Split msg_query and msg_getmset
Originally, `msg_query()` handled two commands by the client: `MSG_QUERY` and
also `MSG_GETMSET` (which was marked as "used during a conversation"). This
changes this and makes it so `msg_query()` returns even when the client hasn't
yet sent a `MSG_GETMSET`, but instead only keeping the current satate of the
query in the `MatchState` struct and setting the expected `required_type` for
the next command.

The purpose of this is that an event based remote protocol server might still
not yet have the `MSG_GETMSET` message from the client as a response, and it'd
be bad practice (or catastrophic) to block the thread while the message arrives.
This blocking (if any) should be handled further up, by the event loop.